Rename _logging_config_updater.py to _logging_actor.py
Also rename the test file name to match the new module name.
